refactor: 버튼 공통 컴포넌트 타입 분리 및 JSDoc 주석 추가 [FE-14]
