+++ shared/components/button/CustomButton.tsx
@@ -1,16 +1,22 @@
-import React from "react";
-import { View, Pressable, StyleSheet, PressableProps } from "react-native";
+import { View, Pressable, StyleSheet } from "react-native";
 import { Label } from "../label/Label";
 import theme from "@/shared/styles/theme";
+import { ICustomButton } from "@/shared/types";
 
-interface ICustomButton extends PressableProps {
-  width?: number;
-  height?: number;
-  radius?: number;
-  icon?: React.ReactNode;
-  disabled?: boolean;
-  children: React.ReactNode;
-}
+/**
+ * 커스텀 버튼 컴포넌트
+ *
+ * @param {number} [width=312] - 버튼의 너비
+ * @param {number} [height=48] - 버튼의 높이
+ * @param {number} [radius=10] - 버튼의 테두리 반경
+ * @param {React.ReactNode} [icon] - 버튼에 표시할 아이콘
+ * @param {boolean} [disabled=false] - 버튼 비활성화 여부
+ * @param {React.ReactNode} children - 버튼에 표시할 텍스트
+ * @param {object} props - 추가적인 Pressable 컴포넌트의 속성
+ * 
+ * @returns {React.ReactElement} - 커스텀 버튼 컴포넌트
+
+ */
 
 export default function CustomButton({
   width = 312,
